Move json code to new json_io module
